add Tradingview

+++ database/migrations/2024_06_02_183813_create_portofolios_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('portofolios', function (Blueprint $table) {
             $table->id();
             $table->foreignId("user_id")->constrained();
-            $table->float("buy_price");
-            $table->float("total_coin");
+            $table->double("buy_price");
+            $table->double("total_coin");
             $table->string("coin_id");
             $table->timestamps();
         });
